RavenDB-13575 : move all ts stuff from QueryResultRetrieverBase to dedicated class TimeSeriesRetriever
